scrapper/requirements.txt
python-dateutil
beautifulsoup4
lxml
requests
